Time series refactored
