--- mango_time_series/mango/utils/__init__.py
@@ -1,2 +1,0 @@
-from mango_time_series.mango.utils.CONST import timedelta_equivalences, PARAMETERS_BASE
-
