+++ mango_time_series/mango_time_series/tests/test_stationary.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 import numpy as np
 import polars as pl
 
-from mango_time_series.mango.time_series.stationary import StationaryTester
+from mango_time_series.time_series.stationary import StationaryTester
 
 
 class TestStationaryTester(TestCase):
