+++ mango_time_series/mango_time_series/tests/test_time_series_dataframe.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from mango_time_series.mango.time_series import TimeSeriesDataFrame
+from mango_time_series.time_series import TimeSeriesDataFrame
 
 
 class TestTimeSeriesDataframe(TestCase):
