+++ mango_time_series/mango_time_series/tests/test_ts_functions.py
@@ -3,7 +3,7 @@ import pandas as pd
 from pandas.testing import assert_frame_equal
 import logging
 
-from mango_time_series.mango.utils.processing import aggregate_to_input_pl
+from mango_time_series.utils.processing import aggregate_to_input_pl
 
 # Define SERIES_CONFIGURATION here for the test
 SERIES_CONF = {
