+++ mango_time_series/mango_time_series/time_series/stationary.py
@@ -6,9 +6,9 @@ import polars as pl
 from statsmodels.tools.sm_exceptions import InterpolationWarning
 from statsmodels.tsa.stattools import adfuller, kpss
 
-from mango_base.mango.logging.logger import get_basic_logger
-from mango_time_series.mango.time_series.decomposition import SeasonalityDecompose
-from mango_time_series.mango.time_series.seasonal import SeasonalityDetector
+from mango.logging.logger import get_basic_logger
+from mango_time_series.time_series.decomposition import SeasonalityDecompose
+from mango_time_series.time_series.seasonal import SeasonalityDetector
 
 logger = get_basic_logger()
