+++ mango_time_series/mango_time_series/utils/__init__.py
@@ -1,0 +1,2 @@
+from mango_time_series.utils.CONST import timedelta_equivalences, PARAMETERS_BASE
+
